+++ src/bundle/toolbar/markup/MToolbarImagePopup.tsx
@@ -3,7 +3,7 @@ import React, {RefObject} from 'react';
 import isNumber from 'is-number';
 
 import {IMG_MAX_HEIGHT, ImageItem, getImageDimensions, insertImages} from '../../../markup';
-import type {CodeEditor} from '../../../markup';
+import type {CodeEditor} from '../../../markup/editor';
 import type {ToolbarBaseProps} from '../../../toolbar';
 import type {UploadSuccessItem} from '../../../utils';
 import {ToolbarImagePopup} from '../custom/ToolbarImagePopup';