+++ src/extensions/additional/GPT/wGptItemData.ts
@@ -1,8 +1,9 @@
 import {cn} from '@bem-react/classname';
 
+import type {WToolbarSingleItemData} from '../../../bundle/toolbar/types';
 import {i18n} from '../../../i18n/gpt/extension';
 import gptIcon from '../../../icons/GPT';
-import {ToolbarDataType, type WToolbarSingleItemData} from '../../../toolbar';
+import {ToolbarDataType} from '../../../toolbar/types';
 
 import {gptHotKeys} from './constants';