+++ src/extensions/markdown/Link/commands.ts
@@ -3,7 +3,7 @@ import type {Command} from 'prosemirror-state';
 
 import {isMarkActive} from '../../../utils/marks';
 
-import {linkType} from '.';
+import {linkType} from './LinkSpecs';
 
 export const removeLink: Command = (state, dispatch) => {
     const linkMarkType = linkType(state.schema);