+++ src/toolbar/Toolbar.tsx
@@ -2,14 +2,14 @@ import React from 'react';
 
 import {cn} from '../classname';
 
-import {ToolbarButtonGroup, ToolbarGroupData} from './ToolbarGroup';
-import {ToolbarBaseProps} from './types';
+import {ToolbarButtonGroup} from './ToolbarGroup';
+import type {ToolbarBaseProps, ToolbarData} from './types';
 
 import './Toolbar.scss';
 
 const b = cn('toolbar');
 
-export type ToolbarData<E> = ToolbarGroupData<E>[];
+export type {ToolbarData};
 
 export type ToolbarProps<E> = ToolbarBaseProps<E> & {
     data: ToolbarData<E>;